Initial import.
Add basic code which is able to parse router descriptor files.  The code
does not support the full specification (according to dir-spec.txt) and
relies on a number of assumptions, i.e., it breaks easily.
Nevertheless, the goal is to be able to quickly analyse historical
router descriptors and Python/Stem is not quite fast enough for that.