Check for file annotations before parsing.
For some assurance regarding what file formats we are dealing with, we
now verify file annotations.  The format of annotations is discussed
here:
<https://collector.torproject.org/formats.html>

If the annotation is not what we expect and support, we abort parsing.
Note that this effectively limits the parser to data formats as archived
by CollecTor.