Add struct representing a network consensus.
Instead of representing a network consensus as a map from string to
RouterStatus, we now use a dedicated struct for this.  On the one hand,
this is handy because we are able to add useful methods to the struct,
for example to do consensus "arithmetic".  On the other hand, we are
able to add meta information such as the publication time of a
consensus.  This is not done yet, though.