Add support for lazy parsing.
If the user is only interested in a subset of a consensus or a bunch of
router descriptors, it is much more efficient to parse on demand rather
than strictly in the beginning.

This commit adds two parsing alternatives which implement lazy parsing.
We leave the decision of whether to use lazy or strict parsing up to the
user.  While in many situations lazy parsing is the way to go, it is
still slower when every single element in a data set is accessed.