Add interfaces for objects and object sets.
We want to be able to run computations over objects (i.e., router
statuses and descriptors) and object sets (i.e., consensuses and
descriptor sets) without knowing what data type we are dealing with.
This commit adds interfaces to do so.  For now, the interface defines:

- Print(): Return the string representation of an object.
- GetFingerprint(): Return the fingerprint of an object.
- Length(): Return the length of an object collection.
- Iterate(): Iterate over the given object collection.
- GetObject(): Return the object identified by the given fingerprint.

This commit renames the ObjectCollector interface to ObjectSet and
removes the PrintObjects() methods as they can now be easily built using
the newly added iterator as well as the Print() method.