Enable object filtering.
When calling an object set's Iterate() method, it now expects an object
filter as parameter.  The object filter can be set to nil, which means
that no filtering is performed.  When an ObjectFilter structure is
given, however, the Iterate() method only returns objects that match the
given filter, e.g., have the correct fingerprint, nickname, or IP
address.