Add a test for extractStatusEntry with lots of initial padding.
DissectFile and its extraction functions have a bug where they may not
return the correct results when the first chunk does not start at the
very beginning of the file.

The extraction functions return a chunk of data, and DissectFile
advances the input by the size of each chunk. But the chunk of data may
not have come from the very beginning of the input. For example, suppose
the extraction function seeks 30 bytes into the input and returns a
10-byte chunk. DissectFile will advance the input by 10 bytes (not 40).
The next call to the extraction function will be called at an input
offset of 10 bytes. The extraction function will seek 20 bytes and find
the same 10-byte chunk that it returned in the previous call.

This may not have been a problem so far, because the chunks tend to be
large in comparison to the initial padding. So the input advancement
will tend to start the next call somewhere in the middle of the
previously returned chunk, preventing the same chunk from being returned
more than once.