Refactor descriptor parsing to work over io.Readers.
The most basic parsing functions now take io.Reader rather than
*os.File. The existing functions such as ParseDescriptorFile have been
rewritten to be in terms of the more basic functions. The lowest-level
functions also decouple type annotation handling from parsing the rest
of the descriptor.

For example, the function ParseDescriptorFile now calls
parseDescriptorFile, which just opens the named file and passes the
resulting io.Reader to parseDescriptor. parseDescriptor reads and checks
the type annotation, then calles parseDescriptorUnchecked ("unchecked"
because it assumes the type annotation has already been checked and
consumed).

ParseUnknownFile is now in terms of a lower-level parseWithAnnotation
that looks at a pre-read Annotation and dispatches to either
parseDescriptorUnchecked or parseConsensusUnchecked.

This change leaves GetAnnotation and CheckAnnotation unused by the other
parts of the API. It adds one new exported function: ParseUnknown
(non-file counterpart to ParseUnknownFile).