Don't fail for the last relay on the status
The bridge authority doesn't add the "directory-signature" at the end of
the file. The scanner function was returning an error when an EOF was
reached and no "directory-signature" is there making the scanner drop
the last relay from the file.

Let's return bufio.ErrFinalToken instead, so the scanner doesn't drop
that record.